+++ src/features/countries/index.ts
@@ -4,6 +4,7 @@ export { CountryDetailsPanel } from "./components/countryDetails/CountryDetailsP
 export { CountryDisplayPanel } from "./components/countryDisplay/CountryDisplayPanel";
 export { CountryFlag } from "./components/countryFlag/CountryFlag";
 export { CountryWithFlag } from "./components/countryFlag/CountryWithFlag";
+export { CountryListGroup } from "./components/countryDetails/CountryListGroup";
 export { CountrySelectModal } from "./components/countrySelect/CountrySelectModal";
 export { CountrySortSelect } from "./components/countrySort/CountrySortSelect";
 export { VisitedStatusIndicator } from "./components/countryDetails/VisitedStatusIndicator";
@@ -44,6 +45,7 @@ export {
 } from "./utils/countryData";
 export {
   filterCountries,
+  filterCountriesByProperty,
   getFilteredIsoCodes,
   getCountryCounts,
   createSovereigntyFilter,