+++ src/features/dashboard/currencies/components/CurrenciesGrid.tsx
@@ -1,0 +1,61 @@
+import React, { useState } from "react";
+import { Link } from "react-router-dom";
+import { EmptyListMessage, SearchInput } from "@components";
+import { type Currency } from "@features/countries";
+import { FaChevronRight } from "react-icons/fa6";
+
+interface CurrenciesGridProps {
+  currencies: Currency[];
+}
+
+export const CurrenciesGrid: React.FC<CurrenciesGridProps> = ({
+  currencies,
+}) => {
+  const [search, setSearch] = useState("");
+  const filtered = currencies.filter(
+    (c) =>
+      c.name.toLowerCase().includes(search.toLowerCase()) ||
+      c.code.toLowerCase().includes(search.toLowerCase()),
+  );
+
+  return (
+    <section className="max-w-6xl mx-auto p-4">
+      <Link
+        to="/dashboard/currencies/exchange"
+        className="flex items-center justify-end text-xl font-semibold mb-4 gap-2 focus:outline-none"
+        aria-label="View full activity log"
+      >
+        <span>Currency Exchange</span>
+        <FaChevronRight className="text-base" />
+      </Link>
+      <div className="flex justify-between items-center mb-4">
+        <SearchInput
+          value={search}
+          onChange={setSearch}
+          placeholder="Search by name or code"
+        />
+      </div>
+      <ul className="divide-y divide-input bg-surface-alt rounded shadow mt-4">
+        {filtered.length === 0 ? (
+          <div className="p-4">
+            <EmptyListMessage message="Currency not found." />
+          </div>
+        ) : (
+          filtered.map((currency) => (
+            <li key={currency.code}>
+              <Link
+                to={`/dashboard/currencies/${currency.code}`}
+                className="grid grid-cols-[100px_1fr] items-center p-4 hover:bg-surface-hover transition rounded focus:outline-none focus:ring-2 focus:ring-primary"
+              >
+                <span className="font-semibold text-lg pr-4">
+                  {currency.code}
+                </span>
+                <span className="text-left">{currency.name}</span>
+              </Link>
+            </li>
+          ))
+        )}
+      </ul>
+    </section>
+  );
+};